requirements.txt
fastapi
uvicorn
httpx
numpy
pillow
tensorflow